moves preprocessor from data/ into src/app/dataValidator
