+++ .env
@@ -1,4 +1,7 @@
-# On running the 'build' commands this file is replaced by the files in `env-config/`
-# To ignore and keep this file we have a command in `postshrinkwrap` which assumes this file is not changing
-# The command is `git update-index --assume-unchanged .env` 
-# Source: https://wildlyinaccurate.com/git-ignore-changes-in-already-tracked-files/
+SIMORGH_BASE_URL=http://localhost:7080
+SIMORGH_ASSETS_MANIFEST_PATH=build/assets-local.json
+SIMORGH_PUBLIC_STATIC_ASSETS_ORIGIN=http://localhost:7080
+SIMORGH_PUBLIC_STATIC_ASSETS_PATH=/
+LOG_DIR='log'
+LOG_LEVEL='debug'
+APP_ENV=local
