+++ src/app/dataValidator/helpers/dataLoader/asyncValidateDir.test.js
@@ -33,7 +33,7 @@ describe('asyncValidateDir helper', () => {
   it('should call readScenario for every file in the /data directory', async () => {
     const readScenarioSpy = jest.spyOn(readScenario, 'readScenario');
 
-    await expectMethodToBeCalledTimes(52, readScenarioSpy);
+    await expectMethodToBeCalledTimes(47, readScenarioSpy);
   });
 
   it('should call fileToValidate for only the valid json file in the /data directory', async () => {
